Refactors game setup logic into new file #5
